Move Custom Include processing into its own module
